# openapi.yaml
openapi: 3.0.0
info:
  title: Cinema API
  description: |
    A comprehensive cinema management API built with NestJS following Domain-Driven Design (DDD) patterns.

    ## Authentication
    This API uses JWT Bearer tokens for authentication. Most endpoints require authentication.

    1. Register a new user at `POST /auth/register`
    2. Login at `POST /auth/login` to receive access and refresh tokens
    3. Include the access token in the `Authorization` header as `Bearer <token>`
    4. When the access token expires, use `POST /auth/refresh` to get new tokens

    ## Permission System
    The API uses a role-based permission system with the following roles:

    | Role | Description |
    |------|-------------|
    | CUSTOMER | Basic user with viewing permissions |
    | BOX_OFFICE_STAFF | Can create bookings and view showtimes |
    | MANAGER | Can manage movies, screens, and showtimes |
    | ADMIN | Full system access |

    ## Error Handling
    All errors follow a consistent response structure with appropriate HTTP status codes.

    ## Pagination
    List endpoints support pagination with `page`, `limit`, `sort_by`, and `sort_order` query parameters.
  version: 1.0.0
  contact:
    name: Cinema API Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from the login endpoint
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - message
        - timestamp
        - path
      properties:
        message:
          type: string
          description: Human-readable error message
          example: "Invalid credentials"
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the error occurred
          example: "2025-12-17T10:30:00.000Z"
        path:
          type: string
          description: The request path that caused the error
          example: "/auth/login"
        errors:
          type: object
          description: Additional error details for validation errors
          additionalProperties: true
          example:
            email: "Invalid email format"
...
    BookingStatus:
      type: string
      enum:
        - PENDING_PAYMENT
        - PAYMENT_PROCESSING
        - CONFIRMED
        - CHECKED_IN
        - CANCELLED
        - EXPIRED
      description: |
        Current status of the booking:
        - PENDING_PAYMENT: Booking created, waiting for payment initiation
        - PAYMENT_PROCESSING: Payment initiated, waiting for confirmation
        - CONFIRMED: Payment received, booking confirmed
        - CHECKED_IN: Customer checked in at the cinema
        - CANCELLED: Booking cancelled by customer or system
        - EXPIRED: Booking expired due to payment timeout
    CreateBookingRequest:
      type: object
      required:
        - showtime_id
        - seat_ids
      properties:
        showtime_id:
          type: string
          pattern: "^SHW_[\\w-]+$"
          maxLength: 100
          description: ID of the showtime to book
          example: "SHW_550e8400-e29b-41d4-a716-446655440000"
        seat_ids:
          type: array
          items:
            type: string
            maxLength: 20
          minItems: 1
          description: List of seat numbers to reserve (e.g., "A1", "A2", "B5")
          example: ["A1", "A2", "A3"]
    CreateBookingResponse:
      type: object
      required:
        - message
        - data
      properties:
        message:
          type: string
          example: "Booking created successfully"
        data:
          type: object
          required:
            - booking_id
            - showtime_id
            - seat_numbers
            - ticket_count
            - total_amount
            - service_fee
            - currency
            - hold_expires_at
            - status
            - created_at
          properties:
            booking_id:
              type: string
              pattern: "^BKG_[\\w-]+$"
              description: Unique booking identifier
              example: "BKG_550e8400-e29b-41d4-a716-446655440000"
            showtime_id:
              type: string
              pattern: "^SHW_[\\w-]+$"
              example: "SHW_550e8400-e29b-41d4-a716-446655440000"
            seat_numbers:
              type: array
              items:
                type: string
              example: ["A1", "A2", "A3"]
            ticket_count:
              type: integer
              minimum: 1
              example: 3
            total_amount:
              type: number
              description: Total amount including service fee (in IDR)
              example: 165000
            service_fee:
              type: number
              description: Service fee (in IDR)
              example: 15000
            currency:
              type: string
              example: "IDR"
            hold_expires_at:
              type: string
              format: date-time
              description: When the hold expires if payment not initiated
              example: "2025-12-18T11:00:00.000Z"
            status:
              $ref: "#/components/schemas/BookingStatus"
            created_at:
              type: string
              format: date-time
              example: "2025-12-18T10:45:00.000Z"
...
    GetBookingResponse:
      type: object
      required:
        - message
        - data
      properties:
        message:
          type: string
          example: "Booking retrieved successfully"
        data:
          type: object
          required:
            - id
            - customer_id
            - showtime_id
            - movie_title
            - screen_name
            - start_time
            - end_time
            - tickets
            - status
            - total_amount
            - service_fee
            - currency
            - payment_url
            - invoice_code
            - created_at
            - hold_expires_at
            - confirmed_at
            - checked_in_at
          properties:
            id:
              type: string
              pattern: "^BKG_[\\w-]+$"
              example: "BKG_550e8400-e29b-41d4-a716-446655440000"
            customer_id:
              type: string
              pattern: "^USR_[\\w-]+$"
              example: "USR_550e8400-e29b-41d4-a716-446655440000"
            showtime_id:
              type: string
              pattern: "^SHW_[\\w-]+$"
              example: "SHW_550e8400-e29b-41d4-a716-446655440000"
            movie_title:
              type: string
              example: "The Matrix Resurrections"
            screen_name:
              type: string
              example: "Screen 1"
            start_time:
              type: string
              format: date-time
              example: "2025-12-20T19:30:00.000Z"
            end_time:
              type: string
              format: date-time
              example: "2025-12-20T21:58:00.000Z"
            tickets:
              type: array
              items:
                $ref: "#/components/schemas/BookingTicket"
            status:
              $ref: "#/components/schemas/BookingStatus"
            total_amount:
              type: number
              example: 165000
            service_fee:
              type: number
              example: 15000
            currency:
              type: string
              example: "IDR"
            payment_url:
              type: string
              nullable: true
              format: uri
              description: Payment gateway URL (available when PENDING_PAYMENT/PAYMENT_PROCESSING)
              example: "https://checkout.xendit.co/invoice/abc123"
            invoice_code:
              type: string
              nullable: true
              description: Invoice code (available after payment)
              example: "INV-2025-001234"
            created_at:
              type: string
              format: date-time
              example: "2025-12-18T10:45:00.000Z"
            hold_expires_at:
              type: string
              format: date-time
              example: "2025-12-18T11:00:00.000Z"
            confirmed_at:
              type: string
              format: date-time
              nullable: true
              description: When the booking was confirmed (payment received)
              example: "2025-12-18T10:50:00.000Z"
            checked_in_at:
              type: string
              format: date-time
              nullable: true
              description: When the customer checked in at the cinema
              example: null
...
    GetAllBookingsResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - items
            - total
            - page
            - limit
            - total_pages
          allOf:
            - $ref: "#/components/schemas/PaginationMeta"
            - type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/BookingListItem"
...
    CancelBookingResponse:
      type: object
      required:
        - message
        - data
      properties:
        message:
          type: string
          example: "Booking cancelled successfully"
        data:
          type: object
          required:
            - booking_id
            - cancelled_at
          properties:
            booking_id:
              type: string
              pattern: "^BKG_[\\w-]+$"
              example: "BKG_550e8400-e29b-41d4-a716-446655440000"
            cancelled_at:
              type: string
              format: date-time
              example: "2025-12-18T10:55:00.000Z"
...
  parameters:
...
    BookingIdPath:
      name: booking_id
      in: path
      required: true
      description: Unique booking identifier
      schema:
        type: string
        pattern: "^BKG_[\\w-]+$"
        maxLength: 100
      example: "BKG_550e8400-e29b-41d4-a716-446655440000"
    PageQuery:
      name: page
      in: query
      description: Page number for pagination (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    LimitQuery:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      example: 10
...
    SortOrderQuery:
      name: sort_order
      in: query
      description: Sort direction
      schema:
        type: string
        enum:
          - asc
          - desc
        default: asc
      example: "asc"
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validation_error:
              summary: Validation Error
              value:
                message: "Validation failed"
                timestamp: "2025-12-17T10:30:00.000Z"
                path: "/movies"
                errors:
                  title: "String must contain at most 100 character(s)"
            invalid_email:
              summary: Invalid Email Format
              value:
                message: "Invalid email format"
                timestamp: "2025-12-17T10:30:00.000Z"
                path: "/auth/register"
            invalid_password:
              summary: Invalid Password Format
              value:
                message: "Password must contain at least one lowercase letter, one uppercase letter, one number, and one special character"
                timestamp: "2025-12-17T10:30:00.000Z"
                path: "/auth/register"
            invalid_id_format:
              summary: Invalid ID Format
              value:
                message: "Invalid movie ID format"
                timestamp: "2025-12-17T10:30:00.000Z"
                path: "/movies/invalid-id"
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalid_credentials:
              summary: Invalid Credentials
              value:
                message: "Invalid email or password"
                timestamp: "2025-12-17T10:30:00.000Z"
                path: "/auth/login"
            invalid_token:
              summary: Invalid JWT Token
              value:
                message: "Invalid or malformed token"
                timestamp: "2025-12-17T10:30:00.000Z"
                path: "/movies"
            expired_token:
              summary: Expired JWT Token
              value:
                message: "Token has expired"
                timestamp: "2025-12-17T10:30:00.000Z"
                path: "/movies"
            token_reuse:
              summary: Token Reuse Detected
              value:
                message: "This refresh token has already been used. Please login again."
                timestamp: "2025-12-17T10:30:00.000Z"
                path: "/auth/refresh"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            message: "You do not have permission to perform this action"
            timestamp: "2025-12-17T10:30:00.000Z"
            path: "/movies"
...
paths:
...
  /bookings:
    get:
      tags:
        - Bookings
      summary: Get user's bookings
      description: |
        Retrieve a paginated list of the current user's bookings.

        **Permissions Required:** VIEW:BOOKING or MANAGE:BOOKING

        **Sorting:**
        - `sort_by`: createdAt, holdExpiresAt, status
        - `sort_order`: asc or desc
      operationId: getUserBookings
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/SortOrderQuery"
        - name: sort_by
          in: query
          description: Field to sort by
          schema:
            type: string
            enum:
              - createdAt
              - holdExpiresAt
              - status
          example: "createdAt"
      responses:
        "200":
          description: List of user's bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAllBookingsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags:
        - Bookings
      summary: Create a new booking
      description: |
        Create a new booking to reserve seats for a showtime.

        **Permissions Required:** CREATE:BOOKING or MANAGE:BOOKING

        **Validation Rules:**
        - `showtime_id`: Must be a valid showtime ID (format: SHW_xxx)
        - `seat_ids`: Array of seat numbers (e.g., ["A1", "A2"]), minimum 1

        **Business Rules:**
        - Seats must be available (not already booked)
        - Creates a temporary hold on the seats
        - Hold expires after 15 minutes if payment not initiated
        - Total amount includes base ticket price + service fee
        - Booking is created with status PENDING_PAYMENT
      operationId: createBooking
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateBookingResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Showtime not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: 'Showtime with ID "SHW_xxx" not found'
                timestamp: "2025-12-18T10:30:00.000Z"
                path: "/bookings"
        "409":
          description: Seats not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: "One or more selected seats are not available"
                timestamp: "2025-12-18T10:30:00.000Z"
                path: "/bookings"
  /bookings/admin/all:
    get:
      tags:
        - Bookings
      summary: Get all bookings (Admin)
      description: |
        Retrieve a paginated list of all bookings in the system.

        **Permissions Required:** VIEW_ALL:BOOKING or MANAGE:BOOKING

        **Filtering:**
        - `user_id`: Filter by customer ID
        - `showtime_id`: Filter by showtime ID
        - `status`: Filter by booking status

        **Sorting:**
        - `sort_by`: createdAt, holdExpiresAt, status
        - `sort_order`: asc or desc
      operationId: getAllBookings
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/SortOrderQuery"
        - name: sort_by
          in: query
          description: Field to sort by
          schema:
            type: string
            enum:
              - createdAt
              - holdExpiresAt
              - status
          example: "createdAt"
        - name: user_id
          in: query
          description: Filter by customer ID
          schema:
            type: string
            pattern: "^USR_[\\w-]+$"
          example: "USR_550e8400-e29b-41d4-a716-446655440000"
        - name: showtime_id
          in: query
          description: Filter by showtime ID
          schema:
            type: string
            pattern: "^SHW_[\\w-]+$"
          example: "SHW_550e8400-e29b-41d4-a716-446655440000"
        - name: status
          in: query
          description: Filter by booking status
          schema:
            $ref: "#/components/schemas/BookingStatus"
      responses:
        "200":
          description: List of all bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAllBookingsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /bookings/{booking_id}:
    get:
      tags:
        - Bookings
      summary: Get booking details
      description: |
        Retrieve details of a specific booking.

        **Permissions Required:** VIEW:BOOKING, VIEW_ALL:BOOKING, or MANAGE:BOOKING

        **Returns:**
        - Full booking details including all tickets
        - Payment URL (if payment pending)
        - Invoice code (if confirmed)
        - Status timestamps
      operationId: getBooking
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/BookingIdPath"
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetBookingResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: 'Booking with ID "BKG_xxx" not found'
                timestamp: "2025-12-18T10:30:00.000Z"
                path: "/bookings/BKG_xxx"
  /bookings/{booking_id}/pay:
    post:
      tags:
        - Bookings
      summary: Initiate payment for booking
      description: |
        Initiate payment process for a pending booking.

        **Permissions Required:** CREATE:BOOKING or MANAGE:BOOKING

        **Business Rules:**
        - Booking must be in PENDING_PAYMENT status
        - Generates a payment URL from the payment gateway (Xendit)
        - Updates booking status to PAYMENT_PROCESSING
        - Payment link expires after 15 minutes

        **Flow:**
        1. Call this endpoint to get payment URL
        2. Redirect user to payment URL
        3. User completes payment on payment gateway
        4. Webhook receives payment confirmation
        5. Booking status updates to CONFIRMED
      operationId: initiatePayment
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/BookingIdPath"
      responses:
        "200":
          description: Payment initiated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitiatePaymentResponse"
        "400":
          description: Booking not in valid state for payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: "Booking is not in PENDING_PAYMENT status"
                timestamp: "2025-12-18T10:30:00.000Z"
                path: "/bookings/BKG_xxx/pay"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /bookings/{booking_id}/cancel:
    post:
      tags:
        - Bookings
      summary: Cancel a booking
      description: |
        Cancel a booking and release reserved seats.

        **Permissions Required:** CREATE:BOOKING or MANAGE:BOOKING

        **Business Rules:**
        - Only bookings in PENDING_PAYMENT, PAYMENT_PROCESSING, or CONFIRMED status can be cancelled
        - Cancelled bookings cannot be reinstated
        - Reserved seats are released back to inventory
        - If payment was completed, refund may be processed separately

        **Valid Cancellation States:**
        - PENDING_PAYMENT → CANCELLED
        - PAYMENT_PROCESSING → CANCELLED
        - CONFIRMED → CANCELLED
      operationId: cancelBooking
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/BookingIdPath"
      responses:
        "200":
          description: Booking cancelled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelBookingResponse"
        "400":
          description: Booking cannot be cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: "Booking in CHECKED_IN status cannot be cancelled"
                timestamp: "2025-12-18T10:30:00.000Z"
                path: "/bookings/BKG_xxx/cancel"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /bookings/{booking_id}/ticket/download:
    get:
      tags:
        - Bookings
      summary: Get ticket download link
      description: |
        Generate a pre-signed download URL for ticket or invoice.

        **Permissions Required:** VIEW:BOOKING or MANAGE:BOOKING

        **Download Types:**
        - `qr_code`: QR code ticket for a specific seat (requires seat_number)
        - `invoice`: PDF invoice for the entire booking

        **Business Rules:**
        - Booking must be in CONFIRMED or CHECKED_IN status
        - Download links expire after 1 hour
        - For QR code, the seat_number parameter is required
      operationId: getTicketDownloadLink
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/BookingIdPath"
        - name: type
          in: query
          required: true
          description: Type of download (qr_code or invoice)
          schema:
            $ref: "#/components/schemas/TicketDownloadType"
        - name: seat_number
          in: query
          required: false
          description: Seat number (required when type is qr_code)
          schema:
            type: string
            maxLength: 20
          example: "A1"
      responses:
        "200":
          description: Download link generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDownloadResponse"
        "400":
          description: Invalid request (e.g., missing seat_number for qr_code)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: "seat_number is required when type is qr_code"
                timestamp: "2025-12-18T10:30:00.000Z"
                path: "/bookings/BKG_xxx/ticket/download"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ==================== Webhook Endpoints ====================
  /webhooks/xendit:
    post:
      tags:
        - Webhooks
      summary: Xendit payment webhook
      description: |
        Handle payment status updates from Xendit payment gateway.

        **Security:**
        - Webhook signature is verified using X-CALLBACK-TOKEN header
        - Invalid signatures return 401 Unauthorized

        **Payment Status Handling:**
        - PAID/SETTLED: Booking confirmed, tickets generated, confirmation email sent
        - EXPIRED: Booking marked as expired, seats released
        - FAILED: Booking cancelled, seats released

        **Note:** This endpoint is called by Xendit, not by the client application.
      operationId: handleXenditWebhook
      parameters:
        - name: x-callback-token
          in: header
          required: true
          description: Xendit callback token for signature verification
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/XenditWebhookPayload"
      responses:
        "200":
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                    example: true
        "401":
          description: Invalid webhook signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: "Invalid webhook signature"
                timestamp: "2025-12-18T10:30:00.000Z"
                path: "/webhooks/xendit"